tricycle primitives
